Added two new featues
Compact mode - an option that saves space as much as possible. Turn off filters as well for maximum real state usage

Changes - This is useful when you have many columns to display but only some of them has values that change. With this option on, you see only the columns that has changes.